+++ client/src/components/Gallery/EditPicture.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import DeletePicture from './DeletePicture';
-import AddPicture from './AddPicture';
 
 export const EditPicture = ({ picture, onCancel, id }) => {
   const [title, setTitle] = useState('');
